move DoNothingAutoencoder to transform init for convenient import
