+++ external/fv3fit/fv3fit/reservoir/transformers/__init__.py
@@ -1,6 +1,6 @@
 from .autoencoder import Autoencoder
 from .sk_transformer import SkTransformer
-from .transformer import encode_columns
+from .transformer import encode_columns, DoNothingAutoencoder
 from typing import Union
 
 ReloadableTransfomer = Union[Autoencoder, SkTransformer]
